+++ ios/Classes/FlutterBluePlusPlugin.m
@@ -66,23 +66,23 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 }
 
 ////////////////////////////////////////////////////////////
-// ██   ██   █████   ███    ██  ██████   ██       ███████
-// ██   ██  ██   ██  ████   ██  ██   ██  ██       ██
-// ███████  ███████  ██ ██  ██  ██   ██  ██       █████
-// ██   ██  ██   ██  ██  ██ ██  ██   ██  ██       ██
-// ██   ██  ██   ██  ██   ████  ██████   ███████  ███████
-//
-// ███    ███  ███████  ████████  ██   ██   ██████   ██████
-// ████  ████  ██          ██     ██   ██  ██    ██  ██   ██
-// ██ ████ ██  █████       ██     ███████  ██    ██  ██   ██
-// ██  ██  ██  ██          ██     ██   ██  ██    ██  ██   ██
-// ██      ██  ███████     ██     ██   ██   ██████   ██████
-//
-//  ██████   █████   ██       ██
-// ██       ██   ██  ██       ██
-// ██       ███████  ██       ██
-// ██       ██   ██  ██       ██
-//  ██████  ██   ██  ███████  ███████
+// ██   ██   █████   ███    ██  ██████   ██       ███████    
+// ██   ██  ██   ██  ████   ██  ██   ██  ██       ██         
+// ███████  ███████  ██ ██  ██  ██   ██  ██       █████      
+// ██   ██  ██   ██  ██  ██ ██  ██   ██  ██       ██         
+// ██   ██  ██   ██  ██   ████  ██████   ███████  ███████                                                       
+//                                                      
+// ███    ███  ███████  ████████  ██   ██   ██████   ██████  
+// ████  ████  ██          ██     ██   ██  ██    ██  ██   ██ 
+// ██ ████ ██  █████       ██     ███████  ██    ██  ██   ██ 
+// ██  ██  ██  ██          ██     ██   ██  ██    ██  ██   ██ 
+// ██      ██  ███████     ██     ██   ██   ██████   ██████                                              
+//                                                      
+//  ██████   █████   ██       ██                           
+// ██       ██   ██  ██       ██                           
+// ██       ███████  ██       ██                           
+// ██       ██   ██  ██       ██                           
+//  ██████  ██   ██  ███████  ███████                     
 
 - (void)handleMethodCall:(FlutterMethodCall *)call result:(FlutterResult)result
 {
@@ -91,7 +91,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
         if (_logLevel >= debug) {
             NSLog(@"[FBP-iOS] handleMethodCall: %@", call.method);
         }
-
+        
         if ([@"setLogLevel" isEqualToString:call.method])
         {
             NSNumber *logLevelIndex = [call arguments];
@@ -106,15 +106,15 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
             };
 
             self.centralManager = [[CBCentralManager alloc] initWithDelegate:self queue:nil options:options];
-
-            NSDictionary *data = [self toBluetoothStateProto:self->_centralManager.state];
-
-            self.stateStreamHandler.cachedBluetoothState = data;
+            
+            NSDictionary *data = [self toBluetoothAdapterStateProto:self->_centralManager.state];
+            
+            self.stateStreamHandler.cachedBluetoothAdapterState = data;
         }
-        if ([@"state" isEqualToString:call.method])
+        if ([@"getAdapterState" isEqualToString:call.method])
         {
-            NSDictionary *data = [self toBluetoothStateProto:self->_centralManager.state];
-
+            NSDictionary *data = [self toBluetoothAdapterStateProto:self->_centralManager.state];
+            
             result(data);
         }
         else if ([@"isAvailable" isEqualToString:call.method])
@@ -163,13 +163,14 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
                 uuids = [uuids arrayByAddingObject:[CBUUID UUIDWithString:u]];
             }
 
-            // options
-            NSDictionary *options = @{
-                CBCentralManagerScanOptionAllowDuplicatesKey: allowDuplicates
-            };
+            // Allow duplicates?
+            NSMutableDictionary<NSString *, id> *scanOpts = [NSMutableDictionary new];
+            if ([allowDuplicates boolValue]) {
+                [scanOpts setObject:[NSNumber numberWithBool:YES] forKey:CBCentralManagerScanOptionAllowDuplicatesKey];
+            }
 
             // Start scanning
-            [self->_centralManager scanForPeripheralsWithServices:uuids options:options];
+            [self->_centralManager scanForPeripheralsWithServices:uuids options:scanOpts];
 
             result(@(true));
         }
@@ -207,7 +208,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
             NSString  *remoteId = args[@"remote_id"];
             bool autoConnect    = args[@"auto_connect"] != 0;
 
-            CBPeripheral *peripheral = nil;
+            CBPeripheral *peripheral = nil; 
             if (peripheral == nil)
             {
                 peripheral = [self findPeripheral:remoteId];
@@ -243,10 +244,10 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
                 // note: use CBConnectPeripheralOptionEnableAutoReconnect constant
                 // when iOS 17 is more widely available
                 [options setObject:@(autoConnect) forKey:@"kCBConnectOptionEnableAutoReconnect"];
-            }
+            } 
 
             [_centralManager connectPeripheral:peripheral options:options];
-
+            
             result(@(true));
         }
         else if ([@"disconnect" isEqualToString:call.method])
@@ -262,22 +263,22 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
             }
 
             [_centralManager cancelPeripheralConnection:peripheral];
-
+            
             result(@(true));
         }
-        else if ([@"deviceState" isEqualToString:call.method])
+        else if ([@"getConnectionState" isEqualToString:call.method])
         {
             // remoteId is passed raw, not in a NSDictionary
             NSString *remoteId = [call arguments];
-
+ 
             CBPeripheral *peripheral = [self findPeripheral:remoteId];
             if (peripheral == nil) {
                 NSString* s = @"peripheral not found. try reconnecting.";
-                result([FlutterError errorWithCode:@"deviceState" message:s details:NULL]);
+                result([FlutterError errorWithCode:@"getConnectionState" message:s details:NULL]);
                 return;
             }
 
-            NSDictionary* response = [self toDeviceStateProto:peripheral state:peripheral.state];
+            NSDictionary* response = [self toConnectionStateProto:peripheral connectionState:peripheral.state];
 
             result(response);
         }
@@ -422,7 +423,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
             NSString  *secondaryServiceUuid = args[@"secondary_service_uuid"];
             NSNumber  *writeType            = args[@"write_type"];
             NSString  *value                = args[@"value"];
-
+            
             // Find peripheral
             CBPeripheral *peripheral = [self findPeripheral:remoteId];
             if (peripheral == nil) {
@@ -461,7 +462,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
                 [_dataWaitingToWriteWithoutResponse setObject:args forKey:remoteId];
                 result(@(true));
                 return;
-            }
+            } 
 
             // Find characteristic
             CBCharacteristic *characteristic = [self locateCharacteristic:characteristicUuid
@@ -469,7 +470,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
                                                                 serviceId:serviceUuid
                                                         secondaryServiceId:secondaryServiceUuid];
 
-
+                                                        
             // Write to characteristic
             [peripheral writeValue:[self convertHexToData:value] forCharacteristic:characteristic type:type];
 
@@ -558,7 +559,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 
             // Set notification value
             [peripheral setNotifyValue:[enable boolValue] forCharacteristic:characteristic];
-
+            
             result(@(true));
         }
         else if ([@"mtu" isEqualToString:call.method])
@@ -583,7 +584,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
                 @"mtu" : @(mtu),
                 @"success" : @(1),
             };
-
+            
             result(response);
         }
         else if ([@"requestMtu" isEqualToString:call.method])
@@ -618,19 +619,19 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
         }
         else if([@"requestConnectionPriority" isEqualToString:call.method])
         {
-            result([FlutterError errorWithCode:@"requestConnectionPriority"
+            result([FlutterError errorWithCode:@"requestConnectionPriority" 
                                     message:@"iOS does not support connection priority requests"
                                     details:NULL]);
         }
         else if([@"setPreferredPhy" isEqualToString:call.method])
         {
-            result([FlutterError errorWithCode:@"setPreferredPhy"
+            result([FlutterError errorWithCode:@"setPreferredPhy" 
                                     message:@"iOS does not support set preferred phy requests"
                                     details:NULL]);
         }
         else if([@"removeBond" isEqualToString:call.method])
         {
-            result([FlutterError errorWithCode:@"removeBond"
+            result([FlutterError errorWithCode:@"removeBond" 
                                     message:@"plugin does not support removeBond function on iOS"
                                     details:NULL]);
         }
@@ -648,17 +649,17 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 }
 
 //////////////////////////////////////////////////////////////////////
-// ██████   ██████   ██  ██    ██   █████   ████████  ███████
-// ██   ██  ██   ██  ██  ██    ██  ██   ██     ██     ██
-// ██████   ██████   ██  ██    ██  ███████     ██     █████
-// ██       ██   ██  ██   ██  ██   ██   ██     ██     ██
+// ██████   ██████   ██  ██    ██   █████   ████████  ███████ 
+// ██   ██  ██   ██  ██  ██    ██  ██   ██     ██     ██      
+// ██████   ██████   ██  ██    ██  ███████     ██     █████   
+// ██       ██   ██  ██   ██  ██   ██   ██     ██     ██      
 // ██       ██   ██  ██    ████    ██   ██     ██     ███████
 //
-// ██    ██  ████████  ██  ██       ███████
-// ██    ██     ██     ██  ██       ██
-// ██    ██     ██     ██  ██       ███████
-// ██    ██     ██     ██  ██            ██
-//  ██████      ██     ██  ███████  ███████
+// ██    ██  ████████  ██  ██       ███████ 
+// ██    ██     ██     ██  ██       ██      
+// ██    ██     ██     ██  ██       ███████ 
+// ██    ██     ██     ██  ██            ██ 
+//  ██████      ██     ██  ███████  ███████ 
 
 - (NSData *)convertHexToData:(NSString *)hexString
 {
@@ -809,36 +810,36 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 
 
 /////////////////////////////////////////////////////////////////////////////////////
-//  ██████  ██████    ██████  ███████  ███    ██  ████████  ██████    █████  ██
-// ██       ██   ██  ██       ██       ████   ██     ██     ██   ██  ██   ██ ██
-// ██       ██████   ██       █████    ██ ██  ██     ██     ██████   ███████ ██
-// ██       ██   ██  ██       ██       ██  ██ ██     ██     ██   ██  ██   ██ ██
-//  ██████  ██████    ██████  ███████  ██   ████     ██     ██   ██  ██   ██ ███████
-//
-// ███    ███   █████   ███    ██   █████    ██████   ███████  ██████
-// ████  ████  ██   ██  ████   ██  ██   ██  ██        ██       ██   ██
-// ██ ████ ██  ███████  ██ ██  ██  ███████  ██   ███  █████    ██████
-// ██  ██  ██  ██   ██  ██  ██ ██  ██   ██  ██    ██  ██       ██   ██
-// ██      ██  ██   ██  ██   ████  ██   ██   ██████   ███████  ██   ██
-//
-// ██████   ███████  ██       ███████   ██████    █████   ████████  ███████
-// ██   ██  ██       ██       ██       ██        ██   ██     ██     ██
-// ██   ██  █████    ██       █████    ██   ███  ███████     ██     █████
-// ██   ██  ██       ██       ██       ██    ██  ██   ██     ██     ██
-// ██████   ███████  ███████  ███████   ██████   ██   ██     ██     ███████
+//  ██████  ██████    ██████  ███████  ███    ██  ████████  ██████    █████  ██      
+// ██       ██   ██  ██       ██       ████   ██     ██     ██   ██  ██   ██ ██      
+// ██       ██████   ██       █████    ██ ██  ██     ██     ██████   ███████ ██      
+// ██       ██   ██  ██       ██       ██  ██ ██     ██     ██   ██  ██   ██ ██      
+//  ██████  ██████    ██████  ███████  ██   ████     ██     ██   ██  ██   ██ ███████ 
+//                                                                                                                                          
+// ███    ███   █████   ███    ██   █████    ██████   ███████  ██████               
+// ████  ████  ██   ██  ████   ██  ██   ██  ██        ██       ██   ██              
+// ██ ████ ██  ███████  ██ ██  ██  ███████  ██   ███  █████    ██████               
+// ██  ██  ██  ██   ██  ██  ██ ██  ██   ██  ██    ██  ██       ██   ██              
+// ██      ██  ██   ██  ██   ████  ██   ██   ██████   ███████  ██   ██              
+//                                                                                                                                                   
+// ██████   ███████  ██       ███████   ██████    █████   ████████  ███████          
+// ██   ██  ██       ██       ██       ██        ██   ██     ██     ██               
+// ██   ██  █████    ██       █████    ██   ███  ███████     ██     █████            
+// ██   ██  ██       ██       ██       ██    ██  ██   ██     ██     ██               
+// ██████   ███████  ███████  ███████   ██████   ██   ██     ██     ███████ 
 
 - (void)centralManagerDidUpdateState:(nonnull CBCentralManager *)central
 {
     if (_logLevel >= debug) {
         NSLog(@"[FBP-iOS] centralManagerDidUpdateState");
     }
-
-    NSDictionary *data = [self toBluetoothStateProto:self->_centralManager.state];
-
+    
+    NSDictionary *data = [self toBluetoothAdapterStateProto:self->_centralManager.state];
+    
     if (_stateStreamHandler.sink != nil) {
         self.stateStreamHandler.sink(data);
     } else {
-        self.stateStreamHandler.cachedBluetoothState = data;
+        self.stateStreamHandler.cachedBluetoothAdapterState = data;
     }
 }
 
@@ -850,7 +851,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     if (_logLevel >= debug) {
         //NSLog(@"[FBP-iOS] centralManager didDiscoverPeripheral");
     }
-
+    
     [self.scannedPeripherals setObject:peripheral forKey:[[peripheral identifier] UUIDString]];
 
     NSDictionary *result = [self toScanResultProto:peripheral advertisementData:advertisementData RSSI:RSSI];
@@ -878,8 +879,8 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     [_channel invokeMethod:@"MtuSize" arguments:response];
 
     // Send connection state
-    [_channel invokeMethod:@"DeviceState"
-                 arguments:[self toDeviceStateProto:peripheral state:peripheral.state]];
+    [_channel invokeMethod:@"getConnectionState"
+                 arguments:[self toConnectionStateProto:peripheral connectionState:peripheral.state]];
 }
 
 - (void)centralManager:(CBCentralManager *)central
@@ -897,8 +898,8 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     peripheral.delegate = nil;
 
     // Send connection state
-    [_channel invokeMethod:@"DeviceState"
-                 arguments:[self toDeviceStateProto:peripheral state:peripheral.state]];
+    [_channel invokeMethod:@"getConnectionState"
+                 arguments:[self toConnectionStateProto:peripheral connectionState:peripheral.state]];
 }
 
 - (void)centralManager:(CBCentralManager *)central
@@ -913,22 +914,22 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     }
 
     // Send connection state
-    [_channel invokeMethod:@"DeviceState"
-                 arguments:[self toDeviceStateProto:peripheral state:peripheral.state]];
+    [_channel invokeMethod:@"getConnectionState"
+                 arguments:[self toConnectionStateProto:peripheral connectionState:peripheral.state]];
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////
-//  ██████  ██████   ██████   ███████  ██████   ██  ██████   ██   ██  ███████  ██████    █████   ██
-// ██       ██   ██  ██   ██  ██       ██   ██  ██  ██   ██  ██   ██  ██       ██   ██  ██   ██  ██
-// ██       ██████   ██████   █████    ██████   ██  ██████   ███████  █████    ██████   ███████  ██
-// ██       ██   ██  ██       ██       ██   ██  ██  ██       ██   ██  ██       ██   ██  ██   ██  ██
+//  ██████  ██████   ██████   ███████  ██████   ██  ██████   ██   ██  ███████  ██████    █████   ██      
+// ██       ██   ██  ██   ██  ██       ██   ██  ██  ██   ██  ██   ██  ██       ██   ██  ██   ██  ██      
+// ██       ██████   ██████   █████    ██████   ██  ██████   ███████  █████    ██████   ███████  ██      
+// ██       ██   ██  ██       ██       ██   ██  ██  ██       ██   ██  ██       ██   ██  ██   ██  ██      
 //  ██████  ██████   ██       ███████  ██   ██  ██  ██       ██   ██  ███████  ██   ██  ██   ██  ███████
 //
-// ██████   ███████  ██       ███████   ██████    █████   ████████  ███████
-// ██   ██  ██       ██       ██       ██        ██   ██     ██     ██
-// ██   ██  █████    ██       █████    ██   ███  ███████     ██     █████
-// ██   ██  ██       ██       ██       ██    ██  ██   ██     ██     ██
-// ██████   ███████  ███████  ███████   ██████   ██   ██     ██     ███████
+// ██████   ███████  ██       ███████   ██████    █████   ████████  ███████          
+// ██   ██  ██       ██       ██       ██        ██   ██     ██     ██               
+// ██   ██  █████    ██       █████    ██   ███  ███████     ██     █████            
+// ██   ██  ██       ██       ██       ██    ██  ██   ██     ██     ██               
+// ██████   ███████  ███████  ███████   ██████   ██   ██     ██     ███████ 
 
 - (void)peripheral:(CBPeripheral *)peripheral didDiscoverServices:(NSError *)error
 {
@@ -1105,8 +1106,8 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     CBDescriptor *cccd = nil;
     if (error == nil) {
         cccd = [self findCCCDescriptor:characteristic];
-        if (cccd == nil) {
-            error = [NSError errorWithDomain:@"com.flutter_blue_plus" code:-1
+        if (cccd == nil) { 
+            error = [NSError errorWithDomain:@"com.flutter_blue_plus" code:-1 
                 userInfo:@{ NSLocalizedDescriptionKey : @"findCCCDescriptor failed"}];
         }
     }
@@ -1127,7 +1128,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
         // Request a read
         // Note: invokeMethod:@"SetNotificationResponse" is invoked later
         // in the didUpdateValueForDescriptor callback
-        [peripheral readValueForDescriptor:cccd];
+        [peripheral readValueForDescriptor:cccd]; 
     }
 }
 
@@ -1164,7 +1165,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     };
 
     int value = [descriptor.value intValue];
-
+    
     // See BmReadDescriptorResponse
     NSDictionary* result = @{
         @"request":         q,
@@ -1261,12 +1262,12 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     if (_logLevel >= debug) {
         NSLog(@"[FBP-iOS] peripheralIsReadyToSendWriteWithoutResponse");
     }
-
+    
     NSDictionary *request = [_dataWaitingToWriteWithoutResponse objectForKey:[[peripheral identifier] UUIDString]];
     if (request == nil) {
         return;
     }
-
+    
     // See BmWriteCharacteristicRequest
     NSString  *characteristicUuid   = request[@"characteristic_uuid"];
     NSString  *serviceUuid          = request[@"service_uuid"];
@@ -1295,20 +1296,20 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 }
 
 //////////////////////////////////////////////////////////////////////
-// ███    ███  ███████   ██████
-// ████  ████  ██       ██
-// ██ ████ ██  ███████  ██   ███
-// ██  ██  ██       ██  ██    ██
-// ██      ██  ███████   ██████
-//
-// ██   ██  ███████  ██       ██████   ███████  ██████   ███████
-// ██   ██  ██       ██       ██   ██  ██       ██   ██  ██
-// ███████  █████    ██       ██████   █████    ██████   ███████
-// ██   ██  ██       ██       ██       ██       ██   ██       ██
-// ██   ██  ███████  ███████  ██       ███████  ██   ██  ███████
+// ███    ███  ███████   ██████      
+// ████  ████  ██       ██           
+// ██ ████ ██  ███████  ██   ███     
+// ██  ██  ██       ██  ██    ██     
+// ██      ██  ███████   ██████ 
+//     
+// ██   ██  ███████  ██       ██████   ███████  ██████   ███████ 
+// ██   ██  ██       ██       ██   ██  ██       ██   ██  ██      
+// ███████  █████    ██       ██████   █████    ██████   ███████ 
+// ██   ██  ██       ██       ██       ██       ██   ██       ██ 
+// ██   ██  ███████  ███████  ██       ███████  ██   ██  ███████ 
 
 
-- (NSString *)convertDataToHex:(NSData *)data
+- (NSString *)convertDataToHex:(NSData *)data 
 {
     const unsigned char *bytes = (const unsigned char *)[data bytes];
     NSMutableString *hexString = [NSMutableString new];
@@ -1320,22 +1321,22 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     return [hexString copy];
 }
 
-- (NSDictionary *)toBluetoothStateProto:(CBManagerState)state
+- (NSDictionary *)toBluetoothAdapterStateProto:(CBManagerState)adapterState
 {
     int value = 0;
-    switch (state)
+    switch (adapterState)
     {
-        case CBManagerStateResetting:    value = 3; break; // BmPowerEnum.turningOn
-        case CBManagerStateUnsupported:  value = 1; break; // BmPowerEnum.unavailable
-        case CBManagerStateUnauthorized: value = 2; break; // BmPowerEnum.unauthorized
-        case CBManagerStatePoweredOff:   value = 6; break; // BmPowerEnum.off
-        case CBManagerStatePoweredOn:    value = 4; break; // BmPowerEnum.on
-        default:                         value = 0; break; // BmPowerEnum.unknown
+        case CBManagerStateResetting:    value = 3; break; // BmAdapterStateEnum.turningOn
+        case CBManagerStateUnsupported:  value = 1; break; // BmAdapterStateEnum.unavailable
+        case CBManagerStateUnauthorized: value = 2; break; // BmAdapterStateEnum.unauthorized
+        case CBManagerStatePoweredOff:   value = 6; break; // BmAdapterStateEnum.off
+        case CBManagerStatePoweredOn:    value = 4; break; // BmAdapterStateEnum.on
+        default:                         value = 0; break; // BmAdapterStateEnum.unknown
     }
 
-    // See BmBluetoothPowerState
+    // See BmBluetoothAdapterState
     return @{
-        @"state" : @(value),
+        @"adapter_state" : @(value),
     };
 }
 
@@ -1353,7 +1354,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     // Manufacturer Data
     NSDictionary* manufDataB = nil;
     if (manufData != nil && manufData.length > 2) {
-
+        
         // first 2 bytes are manufacturerId
         unsigned short manufId = 0;
         [manufData getBytes:&manufId length:2];
@@ -1361,12 +1362,12 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
         // trim off first 2 bytes
         NSData* trimmed = [manufData subdataWithRange:NSMakeRange(2, manufData.length - 2)];
         NSString* hex = [self convertDataToHex:trimmed];
-
+        
         manufDataB = @{
             @(manufId): hex,
         };
     }
-
+    
     // Service Uuids - convert from CBUUID's to UUID strings
     NSArray *serviceUuidsB = nil;
     if (serviceData != nil) {
@@ -1376,7 +1377,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
         }
         serviceUuidsB = [mutable copy];
     }
-
+    
     // Service Data - convert from CBUUID's to UUID strings
     NSDictionary *serviceDataB = nil;
     if (serviceData != nil)
@@ -1398,7 +1399,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
         @"service_uuids":      serviceUuidsB ? serviceUuidsB : [NSNull null],
         @"service_data":       serviceDataB  ? serviceDataB  : [NSNull null],
     };
-
+  
     // See BmScanResult
     return @{
         @"device":             [self toDeviceProto:peripheral],
@@ -1417,11 +1418,11 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     };
 }
 
-- (NSDictionary *)toDeviceStateProto:(CBPeripheral *)peripheral
-                               state:(CBPeripheralState)state
+- (NSDictionary *)toConnectionStateProto:(CBPeripheral *)peripheral
+                     connectionState:(CBPeripheralState)connectionState
 {
     int stateIdx = 0;
-    switch (state)
+    switch (connectionState)
     {
         case CBPeripheralStateDisconnected:  stateIdx = 0; break; // BmConnectionStateEnum.disconnected
         case CBPeripheralStateConnecting:    stateIdx = 1; break; // BmConnectionStateEnum.connecting
@@ -1431,8 +1432,8 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 
     // See BmConnectionStateResponse
     return @{
-        @"remote_id": [[peripheral identifier] UUIDString],
-        @"state":     @(stateIdx),
+        @"remote_id":        [[peripheral identifier] UUIDString],
+        @"connection_state": @(stateIdx),
     };
 }
 
@@ -1470,7 +1471,7 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
     for (CBDescriptor *d in [characteristic descriptors])
     {
         int value = [d.value intValue];
-
+    
         // See: BmBluetoothDescriptor
         NSDictionary* desc = @{
             @"uuid":                   [d.UUID fullUUIDString],
@@ -1526,11 +1527,11 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
 }
 
 //////////////////////////////////////////
-// ██    ██ ████████  ██  ██       ███████
-// ██    ██    ██     ██  ██       ██
-// ██    ██    ██     ██  ██       ███████
-// ██    ██    ██     ██  ██            ██
-//  ██████     ██     ██  ███████  ███████
+// ██    ██ ████████  ██  ██       ███████ 
+// ██    ██    ██     ██  ██       ██      
+// ██    ██    ██     ██  ██       ███████ 
+// ██    ██    ██     ██  ██            ██ 
+//  ██████     ██     ██  ███████  ███████ 
 
 - (void)log:(LogLevel)level
      format:(NSString *)format, ...
@@ -1558,9 +1559,9 @@ typedef NS_ENUM(NSUInteger, LogLevel) {
                               eventSink:(FlutterEventSink)eventSink
 {
     self.sink = eventSink;
-    if (self.cachedBluetoothState != nil)
+    if (self.cachedBluetoothAdapterState != nil)
     {
-        self.sink(self.cachedBluetoothState);
+        self.sink(self.cachedBluetoothAdapterState);
     }
     return nil;
 }